pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sitcom-character-styles"
version = "1.0.0"
description = "MCP server for applying sitcom character sensory logic to image generation prompts"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Dal Marsters", email = "dal@lushy.ai"}
]
keywords = ["mcp", "olog", "categorical", "sitcom", "character", "aesthetics", "prompts"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Multimedia :: Graphics",
]

dependencies = [
    "fastmcp>=0.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
sitcom-character-styles = "sitcom_character_styles.mcp_server.server:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages]
find = {}

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["characters/*/tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers"
